+++ CHANGELOG.md
@@ -21,11 +21,17 @@ and start a new "In Progress" section above it.
 
 ## In progress
 
+
+## 0.99.0
+
+- Support `job_options` in synchronous processing (experimental)
+  (related to [Open-EO/openeo-geopyspark-driver#531](https://github.com/Open-EO/openeo-geopyspark-driver/issues/531), eu-cdse/openeo-cdse-infra#114)
+
 ## 0.98.0
 
 - Add `job_options` argument to `OpenEoBackendImplementation.request_costs()` API.
   It's optional and unused for now, but allows openeo-geopyspark-driver to adapt already.
-  (related to [Open-EO/openeo-geopyspark-driver#531](https://github.com/Open-EO/openeo-geopyspark-driver/issues/531))
+  (related to [Open-EO/openeo-geopyspark-driver#531](https://github.com/Open-EO/openeo-geopyspark-driver/issues/531), eu-cdse/openeo-cdse-infra#114)
 
 ## 0.97.0